generate_PA

Generate a random transport network using preferential attachment.

generate_PA(n_nodes, n_sources, m_dist, zero_appeal = 1, compact = TRUE)

Arguments

n_nodes
Number of (non-source) nodes to generate.

n_sources
Number of source nodes to initialize the network with (has to be at least 1). Note that there is no guarantee all source nodes will become part of the network.

m_dist
The probability distribution to draw the number of inputs for new nodes from. m_dist will be normalized, therefore it does not have to sum up to 1.

zero_appeal
Constant to be added to the nodes' attractiveness.

compact
Whether to remove isolated source nodes.

Value

An edgelist as a dataframe.

Details

This function generates a random scale-free network. It uses the Barabasi-Albert preferential attachment algorithm, slightly modified to allow for directedness and isolated initial nodes.

Examples

generate_PA(20, 5, c(3, 1))

#>    from to
#> 1     0  4
#> 2     0  5
#> 3     1  5
#> 4     1  6
#> 5     5  7
#> 6     2  8
#> 7     2  9
#> 8     0 10
#> 9     1 11
#> 10    1 12
#> 11    8 13
#> 12    3 14
#> 13    1 15
#> 14    3 16
#> 15    2 17
#> 16    1 17
#> 17    0 18
#> 18    0 19
#> 19    3 19
#> 20    8 20
#> 21    1 20
#> 22   12 21
#> 23    3 22
#> 24   13 23
#> 25    3 23