popgen_ibm_mixed

Simulate spread of pathogens on the network using a (very) simple individual-based model.

popgen_ibm_mixed(p_net, ini_dist = NULL)

Arguments

p_net
A popsnetwork object.

ini_dist
Initial distribution of allele frequencies (optional). ini_dist has to be a list containing a vector of node IDs (see popsnetwork) and a matrix of allele frequencies. Note that *any* node pre-set in this way will effectively be treated as a source and hide nodes that are further upstream (see set_allele_freqs).

Value

A new popsnetwork object with allele frequencies set for each node.

Details

This function simulates the change of gene frequencies in a population of pathogens as they spread through the transport network starting at the external sources (see popsnetwork). At each node founder effects are assumed to change composition of the population. This change is simulated by directly drawing from the distribution of genotypes and unfected units, respectively. Note that for numerical reasons this model will not work for small absolute numbers of units (input and transfer rates). It is also strongly recommended to only run this model on networks that have been generated with the "units" method (see popsnetwork).

Examples

# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(150, 100, 200))
ext <- data.frame(node=c("A", "B"), rate=c(300, 100), input=c(1000, 1000))
net <- popsnetwork(el, ext, spread_model="units")

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# run simulation on initialized net
ini_net <- set_allele_freqs(net, ini_freqs)
popgen_ibm_mixed(ini_net)

#> Nodes:
#> 
#> id	infected	input	alleles...
#> A	0.3	1000	0.0966667	0.506667	0.396667
#> B	0.1	1000	1	0	0
#> C	0.228	250	0.912281	0.0877193	0
#> D	0.25	200	0.94	0.06	0
#> 
#> Links:
#> 
#> from	to	rate	infected
#> A	C	150	0.28
#> B	C	100	0.15
#> C	D	200	0.25

# or we can initialize and run in one call
popgen_ibm_mixed(net, ini_freqs)

#> Nodes:
#> 
#> id	infected	input	alleles...
#> A	0.3	1000	0.133333	0.493333	0.373333
#> B	0.1	1000	1	0	0
#> C	0.228	250	0.929825	0.0701754	0
#> D	0.25	200	0.92	0.08	0
#> 
#> Links:
#> 
#> from	to	rate	infected
#> A	C	150	0.28
#> B	C	100	0.15
#> C	D	200	0.25