popsnetwork

Create a popsnetwork object.

popsnetwork(links, external, transmission = 0, decay = -1,
  spread_model = "fluid", checks = FALSE)

Arguments

links
A dataframe describing all edges in the graph as well as transfer rates between them. The first two columns are read as inputs and outputs. If there are only two columns all rates are assumed to be 1.

external
A dataframe describing external inputs into the network. The first column is expected to contain node ids (as indices or factors), the second column specifies the amount of infected material in the input. If there is a third column present it can be used to set overall input rates on the respective nodes (this is relevant for the ibm).

transmission
Rate of infection within nodes (i.e. proportion of uninfected material becoming infected).

decay
The decay of material within nodes. If this parameter has a value in [0, 1) transport rates for the entire network will be rescaled so that sum(output) == sum(input) * (1-decay) in all (non-leaf) nodes.

spread_model
How to model spread of pathogens. With "fluid" the substrate carrying the infection and the pathogen itself is essentially treated as a fluid and rates are calculated deterministically. With "units" infection as well as selection of infected vs. uninfected material at outputs is modelled as a stochastic process on discrete units.

checks
Perform some basic integrity checks on input data (currently looks for cycles and disconnected sub-networks).

Value

A popsnetwork object.

Details

A popsnetwork object stores the nodes and edges making up a food transport network and associated data describing spread of infected material on the network. A network is created from a tabular description of its edges.

Edge lists and node ids

Many functions in rpathsonpaths take edge lists as an argument. An edge list is a data frame with at least two columns. These columns can be either integer vectors which will be interpreted as 0-based node indices or factors. In general where applicable functions will return the same format they received.

Note that formats can not be mixed easily. Function arguments (and data frame columns) have to have the same type. Furthermore nodes in a popsnetwork object generated with integer indices can not be referenced by name.

In general integer indices are considerably faster than factors.

Using popsnetwork

Generating simulated genetic data from a popsnetwork object happens in several steps:

Create the basic network using the constructor popsnetwork.

Set initial allele frequencies for a number of nodes using set_allele_freqs. Note that the initial allele distribution can also be set in the next step, allowing this step to be skipped.

Simulate spread of genetic material through the network with popgen_dirichlet.

Draw samples from the simulated population using draw_isolates.

A worked example is available in the 'overview' vignette.

Examples

# 1) fluid model
# this model is pretty resilient to numerical issues
inp <- c(0L, 0L, 1L, 2L, 3L, 1L, 5L)
outp <- c(1L, 2L, 3L, 3L, 4L, 4L, 2L)
rates <- c(1, 1.5, 0.5, 0.1, 1, 0.1, 0.5)
edgelist <- data.frame(inp, outp, rates)
ext <- data.frame(c(0L, 5L), c(0.5, 0.5))
net <- popsnetwork(edgelist, ext, 0.1)

# 2) units model
# for this model absolute numbers are relevant
# note that whether we use integer or string node ids is completely arbitrary 
# NOT RUN {
# this will produce an error in popsnetwork since node C's output is larger than
# its input (300 > 250)
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(150, 100, 300))
ext <- data.frame(node=c("A", "B"), rate=c(300, 100), input=c(1000, 1000))
net <- popsnetwork(el, ext, spread_model="units")
# }

# if we reduce C's output a bit it works fine
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(150, 100, 200))
ext <- data.frame(node=c("A", "B"), rate=c(300, 100), input=c(1000, 1000))
net <- popsnetwork(el, ext, spread_model="units")