run_popsnet

Create a popsnetwork and run the transport simulation.

run_popsnet(edgelist, ini_input, ini_infd, ini_freqs, n = 1L,
  transmission = 0, decay = -1, theta = 1, spread_model = "units",
  drift_model = "units", checks = FALSE)

Arguments

edgelist
The list of edges as a dataframe, with origin in the first and target in the second column. If a third column is present it is interpreted as transfer rates per edge.

ini_input
Net input rates per source node. Note that source nodes in the network will be automatically detected and will be processed in order of node id. ini_input can either be a single number that will be a repeated for each source node or a list containing one element per source node.

ini_infd
Input rate of infected units per source nodes. Format requirements are identical to those for ini_input.

ini_freqs
Allele frequencies per source node. This can be one out of:

a matrix with one row per root node

a list of initialization objects (each of them containing a vector of node ids and a matrix of frequencies, see set_allele_freqs). In this case n simulations are run per initialization.

n
Number of simulation runs to perform.

transmission
Rate of infection in nodes.

decay
Decay rate in nodes - used to infer transfer rates. If this is negative all transfer rates will be assumed to be 1 (which in most cases is not a very useful value) unless edgelist has a third column (see above). If decay is not negative transfer rates will be inferred assuming preservation of mass with decay at the nodes. Note that the ibm simulation requires output to be smaller or equal than input for all nodes.

theta
The shape parameter for the Dirichlet distribution.

spread_model
Either "fluid" or "units".

drift_model
Either "dirichlet" or "units".

checks
Whether to perform some sanity checks on the graph before simulating (slow).

Value

A list containing the network object(s) produced by the simulation(s) as first and the raw network as the second element.

Details

This function creates a popsnetwork object from the given parameters and runs either the ibm or dirichlet simulation on it.

Examples

# the network
edgelist <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"))
# allele frequencies, 3 alleles x 2 source nodes (A, B)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
# run 
run_popsnet(edgelist, 10, 0.1, freqs)

#> $result
#> $result[[1]]
#> Nodes:
#> 
#> id	infected	input	alleles...
#> A	0	10	0	0	0
#> B	0	10	0	0	0
#> C	0	2	0	0	0
#> D	0	1	0	0	0
#> 
#> Links:
#> 
#> from	to	rate	infected
#> A	C	1	0
#> B	C	1	0
#> C	D	1	0
#> 
#> 
#> $raw
#> Nodes:
#> 
#> id	infected	input	alleles...
#> A	0	10
#> B	0	10
#> C	0	2
#> D	0	1
#> 
#> Links:
#> 
#> from	to	rate	infected
#> A	C	1	0
#> B	C	1	0
#> C	D	1	0
#> 
#> $meta
#> $meta$spread
#> [1] "units"
#> 
#> $meta$drift
#> [1] "units"
#> 
#> $meta$n_inp
#> [1] 1
#> 
#> $meta$n_sources
#> [1] 2
#> 
#> $meta$n_ini
#> [1] 1
#> 
#>